Explicit conversions from long to int in benchmarks (for num_threads)

+++ internal/ceres/block_jacobi_preconditioner_benchmark.cc
@@ -63,7 +63,7 @@ static void BM_BlockSparseJacobiPreconditionerBA(benchmark::State& state) {
   Preconditioner::Options preconditioner_options;
   ContextImpl context;
   preconditioner_options.context = &context;
-  preconditioner_options.num_threads = state.range(0);
+  preconditioner_options.num_threads = static_cast<int>(state.range(0));
   context.EnsureMinimumThreads(preconditioner_options.num_threads);
   BlockSparseJacobiPreconditioner p(preconditioner_options, *jacobian);
 
@@ -91,7 +91,7 @@ static void BM_BlockCRSJacobiPreconditionerBA(benchmark::State& state) {
   Preconditioner::Options preconditioner_options;
   ContextImpl context;
   preconditioner_options.context = &context;
-  preconditioner_options.num_threads = state.range(0);
+  preconditioner_options.num_threads = static_cast<int>(state.range(0));
   context.EnsureMinimumThreads(preconditioner_options.num_threads);
   BlockCRSJacobiPreconditioner p(preconditioner_options, jacobian_crs);
 
@@ -124,7 +124,7 @@ static void BM_BlockSparseJacobiPreconditionerUnstructured(
   Preconditioner::Options preconditioner_options;
   ContextImpl context;
   preconditioner_options.context = &context;
-  preconditioner_options.num_threads = state.range(0);
+  preconditioner_options.num_threads = static_cast<int>(state.range(0));
   context.EnsureMinimumThreads(preconditioner_options.num_threads);
   BlockSparseJacobiPreconditioner p(preconditioner_options, *jacobian);
 
@@ -160,7 +160,7 @@ static void BM_BlockCRSJacobiPreconditionerUnstructured(
   Preconditioner::Options preconditioner_options;
   ContextImpl context;
   preconditioner_options.context = &context;
-  preconditioner_options.num_threads = state.range(0);
+  preconditioner_options.num_threads = static_cast<int>(state.range(0));
   context.EnsureMinimumThreads(preconditioner_options.num_threads);
   BlockCRSJacobiPreconditioner p(preconditioner_options, jacobian_crs);
